Move tryGetQueryMetadata out of helpers
