+++ extensions/ql-vscode/src/variant-analysis/run-remote-query.ts
@@ -3,12 +3,7 @@ import { relative, join, sep, dirname, parse, basename } from "path";
 import { dump, load } from "js-yaml";
 import { copy, writeFile, readFile, mkdirp } from "fs-extra";
 import { dir, tmpName } from "tmp-promise";
-import {
-  askForLanguage,
-  findLanguage,
-  tryGetQueryMetadata,
-  tmpDir,
-} from "../helpers";
+import { askForLanguage, findLanguage, tmpDir } from "../helpers";
 import { getOnDiskWorkspaceFolders } from "../common/vscode/workspace-folders";
 import { Credentials } from "../common/authentication";
 import * as cli from "../codeql-cli/cli";
@@ -40,6 +35,7 @@ import {
   QLPACK_LOCK_FILENAMES,
 } from "../pure/ql";
 import { QueryLanguage } from "../common/query-language";
+import { tryGetQueryMetadata } from "../codeql-cli/query-metadata";
 
 export interface QlPack {
   name: string;
